P control for claw
